조대규 : 슬라임 분할 완료

+++ Project/GameEngineContents/SludgeHeartRoom.h
@@ -13,7 +13,7 @@ public:
 	SludgeHeartRoom& operator=(const SludgeHeartRoom& _Other) = delete;
 	SludgeHeartRoom& operator=(SludgeHeartRoom&& _Other) noexcept = delete;
 
-	void SpawnDividedSludge(float _DividedCount, float4 _SpawnPos);
+	void SpawnDividedSludge(int _DividedCount, float4 _SpawnPos);
 
 protected:
